Refactor Main for easier testing
